Add ability to rename .nfo file to .nfo-orig

+++ headphones/postprocessor.py
@@ -389,6 +389,9 @@ def doPostProcessing(albumid, albumpath, release, tracks, downloaded_track_list,
     if headphones.CLEANUP_FILES:
         cleanupFiles(albumpath)
 
+    if headphones.KEEP_NFO:
+        renameNFO(albumpath)
+        
     if headphones.ADD_ALBUM_ART and artwork:
         addAlbumArt(artwork, albumpath, release)
 
@@ -560,11 +563,23 @@ def cleanupFiles(albumpath):
     for r,d,f in os.walk(albumpath):
         for files in f:
             if not any(files.lower().endswith('.' + x.lower()) for x in headphones.MEDIA_FORMATS):
-                logger.debug('Removing: %s' % files)
-                try:
-                    os.remove(os.path.join(r, files))
+                if not (headphones.KEEP_NFO and files.lower().endswith('.nfo')):
+                    logger.debug('Removing: %s' % files)
+                    try:
+                        os.remove(os.path.join(r, files))
+                    except Exception, e:
+                        logger.error(u'Could not remove file: %s. Error: %s' % (files.decode(headphones.SYS_ENCODING, 'replace'), e))
+
+def renameNFO(albumpath):
+    for r,d,f in os.walk(albumpath):
+        for file in f:
+            if file.lower().endswith('.nfo'):
+                logger.debug('Renaming: "%s" to "%s"' % (file.decode(headphones.SYS_ENCODING, 'replace'), file.decode(headphones.SYS_ENCODING, 'replace') + '-orig'))
+                try:          
+                    new_file_name = os.path.join(r, file)[:-3] + 'orig.nfo'
+                    os.rename(os.path.join(r, file), new_file_name)
                 except Exception, e:
-                    logger.error(u'Could not remove file: %s. Error: %s' % (files.decode(headphones.SYS_ENCODING, 'replace'), e))
+                    logger.error(u'Could not rename file: %s. Error: %s' % (os.path.join(r, file).decode(headphones.SYS_ENCODING, 'replace'), e))
 
 def moveFiles(albumpath, release, tracks):
